+++ sdk/python/feast/infra/aws.py
@@ -6,7 +6,7 @@ from tempfile import TemporaryFile
 from urllib.parse import urlparse
 
 from feast.errors import S3RegistryBucketForbiddenAccess, S3RegistryBucketNotExist
-from feast.infra.provider import PassthroughProvider
+from feast.infra.passthrough_provider import PassthroughProvider
 from feast.protos.feast.core.Registry_pb2 import Registry as RegistryProto
 from feast.registry_store import RegistryStore
 from feast.repo_config import RegistryConfig
