+++ sdk/python/feast/infra/local.py
@@ -7,7 +7,7 @@ import pytz
 
 from feast import FeatureTable
 from feast.feature_view import FeatureView
-from feast.infra.provider import PassthroughProvider
+from feast.infra.passthrough_provider import PassthroughProvider
 from feast.protos.feast.core.Registry_pb2 import Registry as RegistryProto
 from feast.registry_store import RegistryStore
 from feast.repo_config import RegistryConfig
